refactor(deployments): share scoped lookups
